pyproject.toml
[project]
name = "cbr-bank-reports-pipeline"
version = "0.1.0"
description = "High-throughput Celery pipeline for parsing CBR bank reports"
requires-python = ">=3.12"
dependencies = [
  "celery==5.4.0",
  "redis==5.2.1",
  "aiohttp==3.11.16",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.1",
  "orjson==3.10.15",
  "pydantic==2.10.6",
  "python-dateutil==2.9.0.post0",
  "tenacity==9.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
